fix worker iteration for a tree index
Before the patch, during the worker iteration (in case of using
a tree index) an array of tuples for check on expiration was
obtained before suspending, and some tuples can be modified / deleted
from another fiber while the worker fiber is sleeping.
Now an array of tuples for check on expiration obtained just before
checking.