Start key and iterator type
Added the ability from where to start the iterator
and the type of the iterator itself. Start key can
be set as a function (dynamic parameter) or just a
static value. The type of the iterator can be specified
either with the box.index.* constant,
or with the name for example, 'EQ' or box.index.EQ

Needed for: #50