Fix freezes when stopping a task
The problem is that the call to task:stop hangs, because task:stop
sends cancel to the active fibers and waits for these fibers to become
"dead". But there may be some code that can get cancelation using pcall
and skip it, so we have to check the fiber for cancellation.
We do this by checking a flag `worker_canceled` that was previously
only used for atomic iterations. We also fixed the places where this
flag is checked and added comments why in this or that place we
check the cancellation of the worker.

Closes #69