Update documentation and convert to LDoc format
expirationd contains documentation in source code and in a README.
In both places descriptions of API methods are incomplete and
looks sloppy. My experience with LDoc shows that it can help to
place documentation as close to the code as it possible
and generate documentation in HTML format that are looks nice
and convenient for reading.

This change moves all descriptions of API methods to comments in a Lua
source code and removes API descriptions in the README. Descriptions of
API functions contains a types of arguments and these types conforms to
types used by checks module for validation functions arguments. I have
discovered two options that expirationd.start() uses but both were not
documented: `options.vinyl_assumed_space_len` and
`options.vinyl_assumed_space_len_factor`.

Closes #60
Part of #14
Part of #79