test: check_tuples_not_expired_by_timestamp
Transferring the taptest to the luatest testing system with renaming the
tests and using the existing spaces in helpers.

Check that after run tuples is not expired. Let's make sure that after
the expiration starts, no tuples will be archived since the timestamp
has an advantage of 2 seconds.

Updated test's name:

- not expired task -> test_check_tuples_not_expired_by_timestamp

Part of #61