test: shuffle tests
luatest has an option --shuffle that allows to randomise the order of
tests. This can be useful to detect a test that passes just because it
happens to run after an unrelated test that leaves the system in a
favourable state. Patch enable tests shuffling with predefined random
seed specified by $RANDOM. Seed is useful for reproducing a problems
related to the specific test order.