test: fix flaky 'simple expires test'
The test used timings with a resolution of 1 second to remove tuples
by expiration time. Sometimes the timings overlapped and an 'extra'
tuple was removed. An 'extra' from the test point of view.

It is possible to extend the timings resolution and fix the problem.

The patch acts in a different way. It changes the input data so as
not to depend on timings. In addition, the test is now faster.

Closes #90