api: process a task on a writable space by default
expirationd does not process a task by default if a replication is
configured for a node. It does not matter the node is a master or
a replica.

The patch improves the behavior. The behavior becomes different for
RW and RO nodes [1] by default:

* A task processes on all types of spaces for a RW node.
* A task processes on temporary and local spaces for a RO node.

An user can still force the processing using the `force` option.

1. https://www.tarantool.io/en/doc/latest/reference/configuration/#confval-read_only

Closes #42