api: wait until a space or an index is created
After the patch an user can to start expirationd task disregarding
whether a space (and a primary/requested index) is exists or not.
expirationd does not start a worker fiber until the space or index
appears.

Also, after the patch expirationd stops all task fibers if a space
or an index configuration error happens:

* the space or the index disappears (drop/rename);
* an unsupported type of the index;
* an invalid iterator or a start key for the index;

Closes #64
Closes #68
Closes #116